deployment/requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0